pyproject.toml
[build-system]
build-backend = "hatchling.build"
requires = ["hatchling"]

[project]
authors = [
  {"name" = "Adam Gagorik", "email" = "adam.gagorik@gmail.com"}
]
dependencies = [
  "sqlalchemy>=2.0.32",
  "pymysql>=1.1.1",
  "beautifulsoup4>=4.12.3",
  "requests>=2.32.3",
  "pyyaml>=6.0.2"
]
description = "A command line tool for interacting with a FFXI server."
license = {file = "LICENSE"}
name = "ffxiahbot"
readme = "README.md"
requires-python = ">=3.12"
version = "4.1.0"

[project.urls]
Documentation = "https://AdamGagorik.github.io/ffxiahbot/"
Repository = "https://github.com/AdamGagorik/ffxiahbot"

[tool.coverage.report]
skip_empty = true

[tool.coverage.run]
branch = true
source = ["ffxiahbot"]

[tool.deptry.per_rule_ignores]
DEP002 = ["pymysql"]

[tool.mypy]
check_untyped_defs = "True"
disallow_any_unimported = "True"
disallow_untyped_defs = "True"
files = ["ffxiahbot"]
no_implicit_optional = "True"
show_error_codes = "True"
warn_return_any = "True"
warn_unused_ignores = "True"

[tool.pytest.ini_options]
log_cli = true
log_cli_format = "%(message)s"
log_cli_level = "DEBUG"
testpaths = ["tests"]

[tool.ruff]
fix = true
line-length = 120
target-version = "py39"

[tool.ruff.format]
preview = true

[tool.ruff.lint]
ignore = [
  # LineTooLong
  "E501",
  # DoNotAssignLambda
  "E731",
  # Avoid specifying long messages outside the exception class
  "TRY003",
  # Check for execution of untrusted input
  "S603"
]
select = [
  # flake8-2020
  "YTT",
  # flake8-bandit
  "S",
  # flake8-bugbear
  "B",
  # flake8-builtins
  "A",
  # flake8-comprehensions
  "C4",
  # flake8-debugger
  "T10",
  # flake8-simplify
  "SIM",
  # isort
  "I",
  # mccabe
  "C90",
  # pycodestyle
  "E",
  "W",
  # pyflakes
  "F",
  # pygrep-hooks
  "PGH",
  # pyupgrade
  "UP",
  # ruff
  "RUF",
  # tryceratops
  "TRY"
]

[tool.ruff.lint.per-file-ignores]
"tests/*" = ["S101"]

[tool.uv]
dev-dependencies = [
  "mkdocs>=1.6.0",
  "pre-commit>=3.8.0",
  "mypy>=1.11.2",
  "deptry>=0.20.0",
  "pytest>=8.3.2",
  "mkdocs-material>=9.5.33",
  "mkdocstrings[python]>=0.25.2",
  "mkdocs-include-markdown-plugin>=6.2.2",
  "pytest-cov>=5.0.0"
]
